#1764 cleanup in query confs
