+++ apps/statistics/flow-query-conf.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = [
+	require('../../apps-common/query-conf/date-from'),
+	require('../../apps-common/query-conf/date-to'),
+	require('../../apps-common/query-conf/service'),
+	require('../../apps-common/query-conf/mode')
+];
